requirements.txt
astroid==2.0.4
colorama==0.4.0
Django==2.1.2
django-ckeditor==5.6.1
django-js-asset==1.1.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pylint==2.1.1
pytz==2018.6
six==1.11.0
validate-email==1.3
wrapt==1.10.11
